Combine logging utils into one file
